Refactored OnboardingFragment to work with MVVM and DI
