+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/onboarding/OnboardingFragmentViewModel.kt
@@ -2,7 +2,6 @@ package de.rki.coronawarnapp.ui.onboarding
 
 import com.squareup.inject.assisted.AssistedInject
 import de.rki.coronawarnapp.ui.SingleLiveEvent
-import de.rki.coronawarnapp.ui.onboarding.OnboardingNavigationEvents
 import de.rki.coronawarnapp.util.viewmodel.CWAViewModel
 import de.rki.coronawarnapp.util.viewmodel.SimpleCWAViewModelFactory
