+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/ui/onboarding/OnboardingNavigationEvents.kt
@@ -1,4 +1,12 @@
 package de.rki.coronawarnapp.ui.onboarding
 
-class OnboardingNavigationEvents {
+sealed class OnboardingNavigationEvents {
+    object NavigateToOnboardingPrivacy: OnboardingNavigationEvents()
+    object NavigateToEasyLanguageUrl: OnboardingNavigationEvents()
+    object NavigateToOnboardingTracing: OnboardingNavigationEvents()
+    object NavigateToOnboardingFragment: OnboardingNavigationEvents()
+    object NavigateToOnboardingTest: OnboardingNavigationEvents()
+    object ShowCancelDialog: OnboardingNavigationEvents()
+    object NavigateToOnboardingNotifications: OnboardingNavigationEvents()
+    object NavigateToMainActivity: OnboardingNavigationEvents()
 }
