+++ Corona-Warn-App/src/main/res/layout/fragment_onboarding.xml
@@ -2,6 +2,11 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools">
+    <data>
+        <variable
+            name="onboardingFragmentViewModel"
+            type="de.rki.coronawarnapp.ui.onboarding.OnboardingFragmentViewModel" />
+    </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/onboarding_container"
